latest build

+++ src/scalar/2d-fourier-16x16.c
@@ -17,7 +17,7 @@ static inline size_t doz(size_t a, size_t b) {
 void nnp_fft16x16_with_offset__scalar(
 	const float* data,
 	float* transform,
-	uint64_t data_stride, uint64_t transform_stride,
+	size_t data_stride, size_t transform_stride,
 	uint32_t row_count, uint32_t column_count,
 	uint32_t row_offset, uint32_t column_offset)
 {
@@ -172,7 +172,7 @@ void nnp_fft16x16_with_offset__scalar(
 void nnp_ifft16x16_with_offset__scalar(
 	const float* transform,
 	float* data,
-	uint64_t transform_stride, uint64_t data_stride,
+	size_t transform_stride, size_t data_stride,
 	uint32_t row_count, uint32_t column_count,
 	uint32_t row_offset, uint32_t column_offset)
 {
@@ -322,7 +322,7 @@ void nnp_ifft16x16_with_bias__scalar(
 	const float* transform,
 	float* data,
 	const float* bias,
-	uint64_t transform_stride, uint64_t data_stride,
+	size_t transform_stride, size_t data_stride,
 	uint32_t row_count, uint32_t column_count)
 {
 	transform_stride /= sizeof(float);
@@ -473,7 +473,7 @@ void nnp_ifft16x16_with_bias_with_relu__scalar(
 	const float* transform,
 	float* data,
 	const float* bias,
-	uint64_t transform_stride, uint64_t data_stride,
+	size_t transform_stride, size_t data_stride,
 	uint32_t row_count, uint32_t column_count)
 {
 	transform_stride /= sizeof(float);
